Asher Chen commited at 2022-12-24_23-55-52

+++ clear.sh
@@ -1,8 +1,5 @@
 #! /usr/bin/env bash
 # before running this script, please enter `chmod +x clear.sh`
 
-echo "rm -r bin";
-echo "rm -r output";
-
-rm -r bin;
-rm -r output;
+[ -d "bin" ] && rm -r bin && echo "rm -r bin";
+[ -d "output" ] && rm -r output && echo "rm -r output";
